Moved usePrevious; js_only.
