--- assets/src/infrastructure/services/hooks/index.ts
@@ -1,1 +1,0 @@
-export { default as usePrevious } from './usePrevious';
